Move Appbar to components folder
