+++ frontend/src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import './App.css'
 import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom'
-import AppBar from './AppBar/AppBar'
 import Trackerinfo from './Components/Trackerinfo/Trackerinfo'
 import Login from './Components/Login/Login'
 import SearchAndTableComponent from './Components/TrackerSearch/SearchAndTableComponent'
 import ContextProvider from './Context/ContextProvider'
 import LandingPage from './Components/LandingPage/LandingPage'
+import AppBar from './Components/AppBar/AppBar'
 
 function App() {
     return (
